CI-339 Redife logic behind autobould.
In previous commits I tryied to place conf py and all css
to image itself, instead of every repo.

But there is a trouble with multiversion build, which should be
fixed in future.

Current logic is to have "tmpl" in image which could be copied to
spec dir with create_tmpl.sh script.